+++ plusdeck/dbus/config.py
@@ -21,6 +21,11 @@ T = TypeVar("T")
 
 @dataclass
 class StagedAttr(Generic[T]):
+    """
+    A staged attribute. Shows both the active and target value, and how the value is
+    expected to change when applied.
+    """
+
     type: StageType
     active: T
     target: T
@@ -41,6 +46,11 @@ class StagedAttr(Generic[T]):
 
 
 class StagedConfig:
+    """
+    A staged configuration. Shows both the active and target configurations, and how
+    the attributes are expected to change.
+    """
+
     def __init__(self: Self, active_config: Config, target_config: Config) -> None:
         # The configuration currently loaded by the service
         self.active_config: Config = active_config
@@ -49,6 +59,10 @@ class StagedConfig:
         self.dirty = False
 
     def reload_target(self: Self) -> None:
+        """
+        Reload the target config from file.
+        """
+
         self.target_config = Config.from_file(self.file)
         self._check_config_dirty()
 
@@ -67,6 +81,10 @@ class StagedConfig:
         return file
 
     def get(self: Self, name: str) -> StagedAttr[Any]:
+        """
+        Get the staged status of an attribute.
+        """
+
         active_attr = self.active_config.get(name)
         target_attr = self.target_config.get(name)
 
@@ -80,10 +98,18 @@ class StagedConfig:
         return StagedAttr(type=type_, active=active_attr, target=target_attr)
 
     def set(self: Self, name: str, value: str) -> None:
+        """
+        Stage a new value for an attribute.
+        """
+
         self.target_config.set(name, value)
         self._check_attr_dirty(name)
 
     def unset(self: Self, name: str) -> None:
+        """
+        Stage the unsetting of a value for an attribute.
+        """
+
         self.target_config.unset(name)
         self._check_attr_dirty(name)
 
@@ -107,4 +133,8 @@ class StagedConfig:
         )
 
     def to_file(self: Self) -> None:
+        """
+        Save the target config to file.
+        """
+
         self.target_config.to_file()