+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - `plusdeck.dbus.DbusClient` dbus client class
   - `plusdeckd` dbus service CLI
   - `plusdeckctl` dbus client CLI
+- Improved documentation
 
 2025/01/26 Version 2.0.0
 ------------------------